refactor create_xml_corpus with py modules - to be cleaned
